Revert "Send raw history blobs from history service to frontend (#7179)"
This reverts commit e4cfc5af8a952668b6f157688bedcdf4b7bfee29.
