+++ develop/protoc.sh
@@ -37,11 +37,6 @@ done
 color "Update license headers for proto files..."
 go run ./cmd/tools/copyright/licensegen.go --scanDir "$new"
 
-color "Modify history service server interface..."
-sed -i.bak -e \
-    's/GetWorkflowExecutionHistory(context\.Context, \*GetWorkflowExecutionHistoryRequest) (\*GetWorkflowExecutionHistoryResponse, error)/GetWorkflowExecutionHistory(context.Context, *GetWorkflowExecutionHistoryRequest) (*GetWorkflowExecutionHistoryResponseWithRaw, error)/g' \
-    "$new"/temporal/server/api/historyservice/v1/service_grpc.pb.go && rm "$new"/temporal/server/api/historyservice/v1/service_grpc.pb.go.bak
-
 color "Moving proto files into place..."
 old=$api.old
 [[ -d "$api" ]] && mv -f "$api" "$old"
