+++ service/frontend/workflow_handler_test.go
@@ -39,7 +39,6 @@ import (
 	batchpb "go.temporal.io/api/batch/v1"
 	commonpb "go.temporal.io/api/common/v1"
 	enumspb "go.temporal.io/api/enums/v1"
-	failurepb "go.temporal.io/api/failure/v1"
 	filterpb "go.temporal.io/api/filter/v1"
 	historypb "go.temporal.io/api/history/v1"
 	namespacepb "go.temporal.io/api/namespace/v1"
@@ -49,7 +48,6 @@ import (
 	updatepb "go.temporal.io/api/update/v1"
 	workflowpb "go.temporal.io/api/workflow/v1"
 	"go.temporal.io/api/workflowservice/v1"
-	historyspb "go.temporal.io/server/api/history/v1"
 	"go.temporal.io/server/api/historyservice/v1"
 	"go.temporal.io/server/api/historyservicemock/v1"
 	"go.temporal.io/server/api/matchingservice/v1"
@@ -75,7 +73,7 @@ import (
 	"go.temporal.io/server/common/rpc/interceptor"
 	"go.temporal.io/server/common/searchattribute"
 	"go.temporal.io/server/common/tasktoken"
-	"go.temporal.io/server/service/history/tests"
+	e "go.temporal.io/server/service/history/events"
 	"go.temporal.io/server/service/worker/batcher"
 	"go.temporal.io/server/service/worker/scheduler"
 	"go.uber.org/mock/gomock"
@@ -2038,16 +2036,16 @@ func (s *WorkflowHandlerSuite) TestCountWorkflowExecutions() {
 }
 
 func (s *WorkflowHandlerSuite) TestVerifyHistoryIsComplete() {
-	events := make([]*historyspb.StrippedHistoryEvent, 50)
+	events := make([]*historypb.HistoryEvent, 50)
 	for i := 0; i < len(events); i++ {
-		events[i] = &historyspb.StrippedHistoryEvent{EventId: int64(i + 1)}
+		events[i] = &historypb.HistoryEvent{EventId: int64(i + 1)}
 	}
-	var eventsWithHoles []*historyspb.StrippedHistoryEvent
+	var eventsWithHoles []*historypb.HistoryEvent
 	eventsWithHoles = append(eventsWithHoles, events[9:12]...)
 	eventsWithHoles = append(eventsWithHoles, events[20:31]...)
 
 	testCases := []struct {
-		events       []*historyspb.StrippedHistoryEvent
+		events       []*historypb.HistoryEvent
 		firstEventID int64
 		lastEventID  int64
 		isFirstPage  bool
@@ -2081,7 +2079,7 @@ func (s *WorkflowHandlerSuite) TestVerifyHistoryIsComplete() {
 	}
 
 	for i, tc := range testCases {
-		err := persistence.VerifyHistoryIsComplete(tc.events, tc.firstEventID, tc.lastEventID, tc.isFirstPage, tc.isLastPage, tc.pageSize)
+		err := e.VerifyHistoryIsComplete(tc.events, tc.firstEventID, tc.lastEventID, tc.isFirstPage, tc.isLastPage, tc.pageSize)
 		if tc.isResultErr {
 			s.Error(err, "testcase %v failed", i)
 		} else {
@@ -2804,78 +2802,6 @@ func (s *WorkflowHandlerSuite) TestListBatchOperations_InvalidRerquest() {
 	s.ErrorAs(err, &invalidArgumentErr)
 }
 
-func (s *WorkflowHandlerSuite) TestGetWorkflowExecutionHistory_FailedConvertedToContinueAsNew() {
-	config := s.newConfig()
-	wh := s.getWorkflowHandler(config)
-	we := commonpb.WorkflowExecution{WorkflowId: "wid1", RunId: uuid.New().String()}
-	newRunID := uuid.New().String()
-
-	s.mockNamespaceCache.EXPECT().GetNamespaceID(tests.Namespace).Return(tests.NamespaceID, nil).AnyTimes()
-	s.mockNamespaceCache.EXPECT().GetNamespaceName(tests.NamespaceID).Return(tests.Namespace, nil).AnyTimes()
-	s.mockSearchAttributesProvider.EXPECT().GetSearchAttributes(gomock.Any(), gomock.Any()).Return(searchattribute.TestNameTypeMap, nil).AnyTimes()
-
-	req := &workflowservice.GetWorkflowExecutionHistoryRequest{
-		Namespace:              tests.Namespace.String(),
-		Execution:              &we,
-		MaximumPageSize:        10,
-		HistoryEventFilterType: enumspb.HISTORY_EVENT_FILTER_TYPE_CLOSE_EVENT,
-		SkipArchival:           true,
-	}
-	s.mockHistoryClient.EXPECT().GetWorkflowExecutionHistory(gomock.Any(), &historyservice.GetWorkflowExecutionHistoryRequest{
-		NamespaceId: tests.NamespaceID.String(),
-		Request:     req,
-	}).Return(&historyservice.GetWorkflowExecutionHistoryResponse{
-		Response: &workflowservice.GetWorkflowExecutionHistoryResponse{
-			History: &historypb.History{
-				Events: []*historypb.HistoryEvent{
-					{
-						EventId:   int64(5),
-						EventType: enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_FAILED,
-						Attributes: &historypb.HistoryEvent_WorkflowExecutionFailedEventAttributes{
-							WorkflowExecutionFailedEventAttributes: &historypb.WorkflowExecutionFailedEventAttributes{
-								Failure:                      &failurepb.Failure{Message: "this workflow failed"},
-								RetryState:                   enumspb.RETRY_STATE_IN_PROGRESS,
-								WorkflowTaskCompletedEventId: 4,
-								NewExecutionRunId:            newRunID,
-							},
-						},
-					},
-				},
-			},
-		},
-	}, nil).Times(2)
-
-	oldGoSDKVersion := "1.9.1"
-	newGoSDKVersion := "1.10.1"
-
-	// new sdk: should see failed event
-	ctx := headers.SetVersionsForTests(context.Background(), newGoSDKVersion, headers.ClientNameGoSDK, headers.SupportedServerVersions, headers.AllFeatures)
-	resp, err := wh.GetWorkflowExecutionHistory(ctx, req)
-	s.NoError(err)
-	s.False(resp.Archived)
-	event := resp.History.Events[0]
-	s.Equal(int64(5), event.EventId)
-	s.Equal(enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_FAILED, event.EventType)
-	attrs := event.GetWorkflowExecutionFailedEventAttributes()
-	s.Equal("this workflow failed", attrs.Failure.Message)
-	s.Equal(newRunID, attrs.NewExecutionRunId)
-	s.Equal(enumspb.RETRY_STATE_IN_PROGRESS, attrs.RetryState)
-
-	// old sdk: should see continued-as-new event
-	// TODO: We can remove this once we no longer support SDK versions prior to around September 2021.
-	// See comment in workflowHandler.go:GetWorkflowExecutionHistory
-	ctx = headers.SetVersionsForTests(context.Background(), oldGoSDKVersion, headers.ClientNameGoSDK, headers.SupportedServerVersions, "")
-	resp, err = wh.GetWorkflowExecutionHistory(ctx, req)
-	s.NoError(err)
-	s.False(resp.Archived)
-	event = resp.History.Events[0]
-	s.Equal(int64(5), event.EventId)
-	s.Equal(enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_CONTINUED_AS_NEW, event.EventType)
-	attrs2 := event.GetWorkflowExecutionContinuedAsNewEventAttributes()
-	s.Equal(newRunID, attrs2.NewExecutionRunId)
-	s.Equal("this workflow failed", attrs2.Failure.Message)
-}
-
 func (s *WorkflowHandlerSuite) newConfig() *Config {
 	return NewConfig(dc.NewNoopCollection(), numHistoryShards)
 }
@@ -3311,84 +3237,5 @@ func (s *WorkflowHandlerSuite) TestShutdownWorker() {
 	})
 	if err != nil {
 		s.Fail("ShutdownWorker failed:", err)
-	}
-}
-
-// HistoryService returns a different GetWorkflowExecutionHistoryResponse GetWorkflowExecutionHistoryResponseWithRaw to
-// WorkflowHandler. Since HistoryClient is defined with the response type GetWorkflowExecutionHistoryResponse, grpc
-// will deserialize this message to GetWorkflowExecutionHistoryResponse. This is done to avoid the extra CPU usage in
-// history service to deserialize event blobs to []*HistoryEvent. This test ensures that
-// GetWorkflowExecutionHistoryResponseWithRaw is correctly deserialized to GetWorkflowExecutionHistoryResponse.
-func (s *WorkflowHandlerSuite) TestGetWorkflowExecutionHistoryResponseWithRawHistoryEvents() {
-	// Create history events and batches
-	fullHistory := &historypb.History{
-		Events: []*historypb.HistoryEvent{
-			{
-				EventId:   1,
-				EventType: enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_STARTED,
-				Version:   100,
-			},
-			{
-				EventId:   2,
-				EventType: enumspb.EVENT_TYPE_WORKFLOW_TASK_SCHEDULED,
-				Version:   101,
-			},
-			{
-				EventId:   3,
-				EventType: enumspb.EVENT_TYPE_ACTIVITY_TASK_COMPLETED,
-				Version:   102,
-			},
-			{
-				EventId:   4,
-				EventType: enumspb.EVENT_TYPE_TIMER_FIRED,
-				Version:   103,
-			},
-		},
-	}
-
-	batch1 := &historypb.History{
-		Events: fullHistory.Events[:2],
-	}
-
-	batch2 := &historypb.History{
-		Events: fullHistory.Events[2:],
-	}
-
-	// Marshal each batch
-	rawHistory1, err := batch1.Marshal()
-	s.Require().NoError(err)
-
-	db1 := &commonpb.DataBlob{
-		EncodingType: enumspb.ENCODING_TYPE_PROTO3,
-		Data:         rawHistory1,
-	}
-
-	rawHistory2, err := batch2.Marshal()
-	s.Require().NoError(err)
-
-	db2 := &commonpb.DataBlob{
-		EncodingType: enumspb.ENCODING_TYPE_PROTO3,
-		Data:         rawHistory2,
-	}
-
-	// Create a GetWorkflowExecutionHistoryResponse with raw history blobs. This should be wire compatible with
-	// workflowservice.GetWorkflowExecutionHistoryResponse which has repeated HistoryEvent field instead of repeated bytes.
-	rawResp := &historyspb.GetWorkflowExecutionHistoryResponse{
-		History: [][]byte{db1.Data, db2.Data},
-	}
-
-	serializedRawResp, err := rawResp.Marshal()
-	s.Require().NoError(err)
-
-	resp := &workflowservice.GetWorkflowExecutionHistoryResponse{}
-	err = resp.Unmarshal(serializedRawResp)
-	s.Require().NoError(err)
-
-	// Verify resp has same list of history events as fullHistory
-	for i, event := range resp.History.Events {
-		s.Equal(fullHistory.Events[i].EventId, event.EventId)
-		s.Equal(fullHistory.Events[i].Version, event.Version)
-		s.Equal(fullHistory.Events[i].EventType, event.EventType)
-		s.Nil(event.Attributes)
 	}
 }
